# docs/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IBKR Income System \u2014 Web API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/me": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Me",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/nav": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Nav",
        "operationId": "nav_nav_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MeResponse": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "id",
          "role"
        ],
        "title": "MeResponse"
      },
      "NavResponse": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/NavSection"
            },
            "type": "array",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "sections"
        ],
        "title": "NavResponse"
      },
      "NavSection": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "key",
          "label",
          "available"
        ],
        "title": "NavSection"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}